backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
python-dotenv
pydantic
requests
pinecone-client
PyPDF2
python-docx
